refactor: move displayMutationError to errors package
