+++ src/clients/api/queries/getPendingRewards/formatOutput/formatRewardSummaryData.ts
@@ -1,6 +1,6 @@
 import BigNumber from 'bignumber.js';
-import { logError } from 'packages/errors';
 import { VenusLens } from 'packages/contracts';
+import { logError } from 'packages/errors';
 import { Token } from 'types';
 import { convertDollarsToCents, convertMantissaToTokens } from 'utilities';
