+++ src/packages/errors/index.ts
@@ -1,2 +1,3 @@
 export * from './transactionErrors';
+export * from './displayMutationError';
 export * from './logError';
