+++ src/pages/Vai/MintVai/index.tsx
@@ -7,7 +7,7 @@ import {
   LabeledInlineContent,
   Spinner,
 } from 'components';
-import { displayMutationError } from 'errors';
+import { displayMutationError } from 'packages/errors';
 import { useGetToken } from 'packages/tokens';
 import { useTranslation } from 'packages/translations';
 import React, { useCallback, useMemo } from 'react';
